Move amount calculation to the proportions class
