+++ frontend/src/utils/proportions.js
@@ -232,6 +232,42 @@ export default class Proportions {
     );
   }
 
+  calculateAmounts(totalAmount, causes) {
+    const amounts = {};
+
+    causes.data.forEach((cause) => {
+      const causeProportion = this.getProportion(cause.id);
+      cause.attributes.organizations.data.forEach((organization) => {
+        const organizationProportion = this.getSubProportion(
+          cause.id,
+          organization.id,
+        );
+        const proportion = (causeProportion * organizationProportion) / 10000;
+        const amount = totalAmount * proportion;
+        const roundedAmount = Math.round(amount * 100) / 100;
+        if (roundedAmount > 0) {
+          amounts[organization.id] = roundedAmount;
+        }
+      });
+    });
+
+    // Avoid rounding errors
+    const total = Object.values(amounts).reduce((a, b) => a + b, 0);
+    if (total !== totalAmount) {
+      const discrepancy = Math.round((totalAmount - total) * 100) / 100;
+      const timesToAdd = Math.floor(Math.abs(discrepancy) / 0.01);
+      const adder = discrepancy / timesToAdd;
+
+      const keys = Object.keys(amounts);
+      for (let i = 0; i < timesToAdd; i++) {
+        const key = keys.at(-(i % keys.length) - 1);
+        amounts[key] = Math.round((amounts[key] + adder) * 100) / 100;
+      }
+    }
+
+    return amounts;
+  }
+
   static fromStrapiDataWithEqualProportions(data) {
     const equalProportions = data.map(
       (_, i) => Math.floor(100 / data.length) + (i < 100 % data.length ? 1 : 0),
